Fix endpoint authentication method bug
When passing a provider endpoind to manageiq API its authentication
will only save if it includes userid and password or token, but not
both (exclusive or).
This behavior is quite unpredictable, as one would expect and
error or message when passing both options.

This change separates the generation of openshift provider endpoint
from the amazon provider endpoint, thus making sure only one auth
method is used for the provider's endpoint.