Create openshift provider with prometheus endpoint
Added the ability to choose metrics_type: "prometheus" or "hawkular".
This is a new parameter, and it's value is "hawkular" by default for backward compatibility
If metric_type=prometheus then prometheus_hostname and prometheus_port must be defined

Also added a creation example for prometheus in examples/add_openshift_provider.yml
and a test in tests/test_manageiq_provider.py

Temporary hack - No validation for prometheus endpoint authentication until the is supported by manageiq.
This is also stated in the code in a todo